move flagtypes to point of use
