+++ cmd/openshift-sdn/openshift-sdn.go
@@ -1,13 +1,16 @@
 package main
 
 import (
+	goflag "flag"
 	"math/rand"
 	"os"
 	"time"
 
+	"github.com/spf13/pflag"
+
+	utilflag "k8s.io/component-base/cli/flag"
 	"k8s.io/component-base/logs"
 
-	"github.com/openshift/origin/pkg/cmd/flagtypes"
 	"github.com/openshift/origin/pkg/cmd/openshift-sdn"
 )
 
@@ -17,8 +20,10 @@ func main() {
 
 	rand.Seed(time.Now().UTC().UnixNano())
 
+	pflag.CommandLine.SetNormalizeFunc(utilflag.WordSepNormalizeFunc)
+	pflag.CommandLine.AddGoFlagSet(goflag.CommandLine)
+
 	cmd := openshift_sdn.NewOpenShiftSDNCommand("openshift-sdn", os.Stderr)
-	flagtypes.GLog(cmd.PersistentFlags())
 
 	if err := cmd.Execute(); err != nil {
 		os.Exit(1)
