+++ hack/import-restrictions.json
@@ -451,7 +451,6 @@
       "github.com/openshift/origin/pkg/authorization/apis/authorization/install",
       "github.com/openshift/origin/pkg/build/apis/build",
       "github.com/openshift/origin/pkg/build/apis/build/install",
-      "github.com/openshift/origin/pkg/cmd/flagtypes",
       "github.com/openshift/origin/pkg/cmd/templates",
       "github.com/openshift/origin/pkg/image/apis/image",
       "github.com/openshift/origin/pkg/oauth/util",
