--- pkg/cmd/flagtypes/doc.go
@@ -1,3 +1,0 @@
-// Package flagtypes provides types that implement the pflags.Value interface for
-// converting command line flags to objects.
-package flagtypes
