+++ pkg/oc/cli/login/login.go
@@ -16,8 +16,8 @@ import (
 	"k8s.io/kubernetes/pkg/kubectl/util/templates"
 	"k8s.io/kubernetes/pkg/kubectl/util/term"
 
+	"github.com/openshift/oc/pkg/helpers/flagtypes"
 	kubeconfiglib "github.com/openshift/oc/pkg/helpers/kubeconfig"
-	"github.com/openshift/origin/pkg/cmd/flagtypes"
 )
 
 var (
