+++ staging/src/github.com/openshift/oc/pkg/helpers/flagtypes/addr_test.go
